Moves DRFAuthenticatedGraphQLView out of urls.py to a views.py file.
